task2/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
clickhouse-driver==0.2.6
